+++ containers/src/Data/Sequence/Internal.hs
@@ -208,7 +208,7 @@ import Prelude ()
 import Control.Applicative ((<$>), (<**>),  Alternative,
                             liftA3)
 import qualified Control.Applicative as Applicative
-import Control.DeepSeq (NFData(rnf))
+import Control.DeepSeq (NFData(rnf),NFData1(liftRnf))
 import Control.Monad (MonadPlus(..))
 import Data.Monoid (Monoid(..))
 import Data.Functor (Functor(..))
@@ -517,6 +517,10 @@ instance Traversable Seq where
 
 instance NFData a => NFData (Seq a) where
     rnf (Seq xs) = rnf xs
+
+-- | @since 0.7.1
+instance NFData1 Seq where
+    liftRnf rnfx (Seq xs) = liftRnf (liftRnf rnfx) xs
 
 instance Monad Seq where
     return = pure
@@ -1227,6 +1231,12 @@ instance NFData a => NFData (FingerTree a) where
     rnf (Single x) = rnf x
     rnf (Deep _ pr m sf) = rnf pr `seq` rnf sf `seq` rnf m
 
+-- | @since 0.7.1
+instance NFData1 FingerTree where
+    liftRnf _ EmptyT = ()
+    liftRnf rnfx (Single x) = rnfx x
+    liftRnf rnfx (Deep _ pr m sf) = liftRnf rnfx pr `seq` liftRnf (liftRnf rnfx) m `seq` liftRnf rnfx sf
+
 {-# INLINE deep #-}
 deep            :: Sized a => Digit a -> FingerTree (Node a) -> Digit a -> FingerTree a
 deep pr m sf    =  Deep (size pr + size m + size sf) pr m sf
@@ -1329,6 +1339,13 @@ instance NFData a => NFData (Digit a) where
     rnf (Three a b c) = rnf a `seq` rnf b `seq` rnf c
     rnf (Four a b c d) = rnf a `seq` rnf b `seq` rnf c `seq` rnf d
 
+-- | @since 0.7.1
+instance NFData1 Digit where
+    liftRnf rnfx (One a) = rnfx a
+    liftRnf rnfx (Two a b) = rnfx a `seq` rnfx b
+    liftRnf rnfx (Three a b c) = rnfx a `seq` rnfx b `seq` rnfx c
+    liftRnf rnfx (Four a b c d) = rnfx a `seq` rnfx b `seq` rnfx c `seq` rnfx d
+
 instance Sized a => Sized (Digit a) where
     {-# INLINE size #-}
     size = foldl1 (+) . fmap size
@@ -1407,6 +1424,11 @@ instance NFData a => NFData (Node a) where
     rnf (Node2 _ a b) = rnf a `seq` rnf b
     rnf (Node3 _ a b c) = rnf a `seq` rnf b `seq` rnf c
 
+-- | @since 0.7.1
+instance NFData1 Node where
+    liftRnf rnfx (Node2 _ a b) = rnfx a `seq` rnfx b
+    liftRnf rnfx (Node3 _ a b c) = rnfx a `seq` rnfx b `seq` rnfx c
+
 instance Sized (Node a) where
     size (Node2 v _ _)      = v
     size (Node3 v _ _ _)    = v
@@ -1466,6 +1488,10 @@ instance Traversable Elem where
 
 instance NFData a => NFData (Elem a) where
     rnf (Elem x) = rnf x
+
+-- | @since 0.7.1
+instance NFData1 Elem where
+    liftRnf rnfx (Elem x) = rnfx x
 
 -------------------------------------------------------
 -- Applicative construction