+++ containers/src/Data/Tree.hs
@@ -61,7 +61,7 @@ import Control.Monad (liftM)
 import Control.Monad.Fix (MonadFix (..), fix)
 import Data.Sequence (Seq, empty, singleton, (<|), (|>), fromList,
             ViewL(..), ViewR(..), viewl, viewr)
-import Control.DeepSeq (NFData(rnf))
+import Control.DeepSeq (NFData(rnf),NFData1(liftRnf))
 
 #ifdef __GLASGOW_HASKELL__
 import Data.Data (Data)
@@ -300,6 +300,12 @@ foldlMap1 f g =  -- Use a lambda to allow inlining with two arguments
 
 instance NFData a => NFData (Tree a) where
     rnf (Node x ts) = rnf x `seq` rnf ts
+
+-- | @since 0.7.1
+instance NFData1 Tree where
+    liftRnf rnfx = go
+      where
+      go (Node x ts) = rnfx x `seq` liftRnf go ts
 
 -- | @since 0.5.10.1
 instance MonadZip Tree where